Support utf8mb4
When running `create_migrations_table()` on a database with a utf8mb4 default collation, it fails with
```
DBD::mysql::db do failed: Specified key was too long; max key length is 767 bytes at /usr/local/share/perl5/DB/SQL/Migrations.pm line 120.
```

This is because the length is primary key is too long, VARCHAR(255) with 4-byte characters = 1020 bytes, longer than the maximum length for a primary key, which is 767 bytes.

The fix is to enforce `latin1` charset on the column with 1 byte per character. to ensure the length of the primary key is only 255 bytes.